수정

+++ backend/src/main/java/com/iot2ndproject/mobilityhub/domain/vehicle/repository/CarRepository.java
@@ -5,7 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface CarRepository extends JpaRepository<CarEntity,Integer> {
-    Optional<CarEntity> findByCarNumber(String carNumber );
-
+public interface CarRepository extends JpaRepository<CarEntity,Long> {
+    Optional<CarEntity> findByCarNumber(String carNumber);
 }
